+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
  - Renamed Slider min and max property to minimum and maximum
  - The graphics rendering backend was changed to use femtovg or Qt
  - Renamed `initial-focus` to `forward-focus` as focus related property on elements
+ - The "align-" prefix was removed from `TextHorizontalAlignment` and `TextVerticalAlignment`. Either
+   change `align-left` to `left` or qualify with `TextHorizontalAlignment.left`.
 
 ### Added
  - title property to the Window element